Bug 1044878: Editing duplicated content
Added a new feature that the user has to choose yes to
start editing copy. This will force users to interact with the
form before editing with the choice that they are editing
all copies of this content.

When selecting to use content from another note, the
form will update based on their choice. The editor will
remain open if yes has been selected. The editor will
be hidden and only the content will show if 'Yes' has
not be selected.

This is a UI element only - no decision is saved to the
DB as this is an editor based decision.

behatnotneeded

Change-Id: If44522c00672b296242893f0c876e7ff37561dd2