Bug #1824044 'skip to main content' link skips the new page header on collections
Put the pageactions.tpl inside pageheader.tpl so the ordering of
elements is better for screenreaders.

Skip to main content now goes either to #main or #header-content if dealing
with a custom page header.  The collection navigation custom-dropdown is now
focusable, with clicking func and tab to and press enter for accessibility.
Style guide updated accordingly.

behatnotneeded

Change-Id: I54c17a69c3f55aa4f896ea02be01e04220d4dd4c