Bug 1832308: Group setting and isolated institutions
The public setting on a group should not affect
the jointype.

Bring back original 18.10 functionality for
group settings page by:
* Removing the friends customisation.
* Removing affect of Public setting on other toggles.

Bring back original 18.10 functionality for who
can join a group:
* Users will be able to join the group on their own -
  if it is open - regardless of the isolatedinstitutions
  and public settings.

Please note for isolatedinstitutions:
Only site admins can create public groups.
It is up to the site admin to add users to groups
if they are set to public.
It is also up to the site admin to correctly set the jointype
for the group.
Therefore, for open groups, users can join the
group if it's in their institution.

Sponsored by The Australian National University

behatnotneeded

Change-Id: I0901f1ddf3b1c982dab0b3f4725755b6b0cdf5f2