Bug 1827812: Ensure people receive notifications in their preferred language
'Preferred language' is the language they set in their account settings.

It does not mean that text that is written in a different language is
translated automatically. Only Mahara provided language strings are
of concern here.

Note: Preferred language can be changed if user changes language via
the language switcher before logging in

behatnotneeded

Change-Id: I9eebd522bfc222c0dbfa87c5ca5217f4a50d402b