Bug 1813987: Translate old views to new gridstack layout on edit
Old layout pages will be displayed as they were
with the old layout.

New pages are created with the new grid layout.

If we try to edit a page with an old layout,
the page will be translated into the new grid layout.

To translate from an old layout to a new grid layout,
we respect the row, column and order of blocks inside each cell.

All blocks in the same row in old layout will be together
in the new layout, and the first block of each column
of the same row are displayed at the same level to each other
in the new layout.

The blocks will respect the column order
they had in the old layout.

A column of the old layout can have many blocks. In the
new grid layout they respect the same order they had.

As we have 12 columns in the new grid layout,
the blocks that where included in a 5 column row,
will be displayed in smaller columns in the new layout.

TODO:
confirmation before trnaslating a page to the new grid layout

behatnotneeded

Change-Id: I46f5cdbacb026519631503437c10d33f1b2800e5