Bug 1827811: Notes block
do not show note contents when displayed in a modal.
if comments not allowed but license metadata is on, display details link and open model with license data.
make small change to setting of licensemetadata smarty variable
get rid of comments/details links and use the new block header instead.

Change-Id: I9183292278d06a3766ae08b0376237fda8d7206c
Signed-off-by: Lisa Seeto <lisaseeto@catalyst.net.nz>