Bug 1813987: Copying a page (in old or new layout) creates a page with new layout
- Copy a page when original view is using
grid layout
- Copy a page when original view is using
old layout: Translates the original page layout
without modifying it,
and creates a new page with new grid layout
The blocks in the new page will not have height
assigned. The height of the blocks is updated
on first load of the page in the browser.
This can happen when editing the page or displaying it.
- Copy a collection with old layout: creates the new collection
with new grid layout pages. Each page needs to be loaded on the
browser to update the block heights

behatnotneeded

Change-Id: I43b5f9050be192c057cb1803dcd69575c137a9cc