Bug 1813987: Importing views with grid layout
Importing views and collection with new grid layout

- leap2A files containing views/collections with
new grid layout are imported and created with the new layout

- leap2A files containing collections/view with old layout
will be created with a grid layout. This new layout is
the translation of the old layout they had in the leap2A file.
The heights of the blocks will be updated when the
views are first displayed in the browser.

behatnotneeded

Change-Id: Ia43b7a5c11109023d3caa65ba16e3b6aebcdbe75