Bug 1837194: Allow going directly to a message from a URL
For normal messages, going to the inbox and unhiding a message is fine,
but if the message is one that will have been synced to an LMS, the user
is quite likely to come from the LMS via a direct link to the message itself.

This allows us to expand a message - albeit on the proviso that it is on
the first page of results, though this is a scenario that never came up in
a year of user testing.

I did include a Behat test, though I had to write a new step to cover the
specifics of 'going directly to a message'.

Change-Id: I0e68fcc37f1ed453b6c43ce939d8c2a3a60b292e