Bug 1813987: User should confirm before the view is translated
there is an account preference we can set for showing
a warning message before translating the layout of an old view.

By default it's true. It can be set to false in the account preferences
or by selecting the option when the message is displayed.
if this preference is set to
true: if we want to edit a view with old layout,
a message will be displayed to confirm we want to translate the layout.
false: The view gets translated on edit without asking the user.

The message has the options to accept, cancel  or don't ask again.

cancel: redirects to the displayed view
don't ask again: sets account preference to false but can be reverted
in account preferences.
accept: will redirect to the view edit page with a parameter
saying the user accepted the translation

behatnotneeded

Change-Id: I84e1a86d3bb494a1955da9cbfbde996b555bb132