Bug 1839499: Site pages and collections, Site unavailable error after enabling lti plugin
Fixing the issue of sql exception on pages & collections page after logging in as a non admin
user on enabling lti_assessment plugin. This is caused by calling a select statement for
lti_assessmentid column with aggregate function and group by condition,
but missing group by condition for lti_assessmentid.

Change-Id: I0a040018fadcf543452b24174964e929f286e552