Bug 1813987: Accessibility settings and layout
- Added an option to set the profile to accessible.
This can be set in account preferences and
enables one extra field in the settings
of a view, in the basic settings section, to make
the layout accessible.

- by default, if the user account is accessible then
the pages will be crated as accessible.
To have a non accessible page, the option un view settings
needs to be set as 'No'

- Accessible layout only allows to add blocks with width=12
that is the same with of the grid.
This makes the blocks show as a sequence one on top of the other
like a list. The user can only reorder the position
they have in the list.

- When a page is accessible, the adding a block
by drag & drop is disabled

- Reordering of blocks is accessible by keyborad

- The 'add block' button is reachable by keyboard

- add an intro to the page for the screen reader
to explain how the page works

- add identification to each block for screen reader,
blocktype and title (if it’s set)

TODO
* When adding a block to the bottom or the top of the page,
the page should be scrolled to that block

* Floating menu: should have a way to be recognized as a menu
by the screen reader. Check the access keys for the menu.

behatnotneeded

Change-Id: I08417f0f11d747a67900c88c2f675ef5f85b7499