Bug #1833348 Heading on tags page can't be selected
On My tags page, the link next to "Search results for" is not
clickable due to the Edit this tag button floating on top of it
in a transparent layer. Fix give the button group a margin-left
to move it away from the search title. Both the link and the edit
button (sometimes doesn't display) should now be clickable.

behatnotneeded

Change-Id: Idecfbaaf8965b1908a8b6edd3c3300b36da33929