Bug 1813987: many fixes for gridstack
- set min height for grid only in edit page,
so it won't show a lot of empty space for pages
with few blocks
- make the journal block resize when expanding
content
- [FOR TESTING!!] resize the block everytime
a collapsible is expanded or collapsed
inside the block. Need to check what happens
to the layout structure of the blocks that
are under the block that resizes.

behatnotneeded

Change-Id: I64925c1cb518af30f569ac2bbf69baea6c07ec05
(cherry picked from commit a8b9bfb186ed721ccca66d0578f1ae6ea408ba59)