Bug 1850578: Add Tinymce mobile theme
This is required for the editor to show on Android devices

Also hid the image button in the mobile theme as it can't be configured to connect
to our imagebrowser. Tinymce doesn't currently support custom
plugins for the mobile theme. If this changes in future, we
could look at putting it back.

Change-Id: I41f45ffaad1e20b712f59eac7d7e92f00915f394