Bug 1856260: Delete elements when deleting the parent (standard & element)
- When deleting a standard, then delete all the elements that have it as parent
- When deleting an element, then delete all the elements that have it as parent
- it works when multiple elements depend on a deleted parent
- it gets called recursively: if an element was deleted because its parent
was deleted, then delete all the elements also depending on it

- The message displayed when removing one node of the form is hardcoded in the library
I've changed to
"Are you sure you want to remove this item? If there is any other item depending on it, it will be removed too."
since it's had to see if we are deleting a standard or a standard element.

behatnotneeded
Change-Id: Ifc7d7e0e491c398262bd603ea8f348770a952916