Bug 1857382: Removing page description tinymce editor
and replace the editor with a textarea to edit the description

For views already on the site that have descriptions:
- If they can be transform to a simple text, then remove the html
tags from the view description to make it a simple text and
update the DB with it.
- If the description has tables, lists, etc, then create
a text block with that description and place it above
the other blocks of the view.

upgrade: only remove description for new layout pages
and the description must be:
- more than 160 characters long
or
- have styling tags other than <p>, <em>, <br>

translating pages: if the description makes one of the
two conditions true, then create a new block at the top
of the page with it. If not, if it's a simple short text
then leave it as the view description.

import:
- for new gridstaxc layout pages, add a block at the top if needed.
- for old layout, the page will be translated. After this, we
add a block if needed

copy a page:
- if it has new gridstack layout, it was fixed by the upgrade,
so we don't need to do anything
- if it has old layout, the new view will be translated and
we add a new text bloxk at the top with the description if needed

CHECK what to do
- the site templates, Dashboard and profile pages have descriptions
So do the copies for each user.
But those description don't show when viewing a portfolio or a dashboard.
And the user can't change the settings for those pages either.
old pages not translated: display a new block at the top of the page.
There's no change in the layout on the db, there will be a mock block
at the top of the page

behatnotneeded
Change-Id: I7646405a786e823ae486cc325c30b7a713deccee