Bug 1857935 Share dropdown styling
Fix now forces the dropdown box to be wider to fit in the longest
unwrapped line, and has margin on left to align all text away from
usericon, and a margin on right to stop horizontal scroll. The name
is regular weight, and the rest of the info are in a lighter weight.
- added in stylings for text for secondary information (email etc...)

behatnotneeded

Change-Id: I852b4f9364e8932064029d44652dd191c38926cf