Bug 1870444: Cover images for portfolios
- added cove mage field to view/collection settings
- it works for personal/group/institution/site portfolios
- creates the "Cover Images" folder if it doesn't exist to store the cover images
as default location. This is created when navgating to a view/collection settings section.
- there is an upgrade to add 'coverimage' column to view and
collection tables on the db.
- the cover image is displayed in the view/collection card
on pages and collections section
- copy views/collections should copy the cover image
- display the image in the card of pages and collections if it has one.
If there's no image it should display the desctiption or nothing if there's
no description. If it has cover image and a description, then it displays
the image and when hover over it will display the description.
On mobile it will display the description when touched

TODO
- accessibility
- importing views/collections should import the cover image

behatnotneeded
Change-Id: I72e305ba437ba20df23f8ede224d399eb97f02d1