Behat: downgrade composer packages
there are new versions of a few packages and composer
will try to upgrade them before running behat tests.

But these new versions don't work together very well,
so we ned to revert to the previous versions.

To see this changes here is the link to the first
behat run that broke in jenkins
https://test.mahara.org/job/mahara-gerrit/10891/console

this is what was installed by composer
Package operations: 0 installs, 5 updates, 0 removals
  - Updating symfony/css-selector (v3.4.38 => v5.0.5): Downloading (100%)
  - Updating behat/mink (v1.7.1 => v1.8.1): Downloading (100%)
  - Updating behat/mink-selenium2-driver (v1.3.1 => v1.4.0): Downloading (100%)
  - Updating fabpot/goutte (v3.2.3 => v3.3.0): Downloading (100%)
  - Updating behat/mink-browserkit-driver (1.3.3 => v1.3.4): Downloading (100%)

Note: If it still doesn't work for you in local machine,
remove the content of the behat folder on the mahara dataroot directory

behatnotneeded

Change-Id: Ibd48275c297a44527d551692bee1e4c4294d302b