Bug 1857936: Lock page and artefact instructions
- A new page setting “Template” is added to the
“Advanced” panel on all portfolio page areas of
Mahara (personal, group, institution, site)
with a Yes/No switch
- added 2 columns to view table:
'templatetocopy': indicates if the view is a
  template that can be copied
'originaltemplate': if it has a value then the view
  is a copy of a template and should have
  the instructions locked.
  The value it contains is the viewid of
  the original template.
- if a view is a copy of a template: display
instructions in view settings and text blocks
but don't allow to edit them. If empty, don't show them.
- When adding a new text block to a page,
don't allow to add instructions
- Allow to remove the lock on a copy for roles:
    site admins in institution and site pages
    institution admins in institution pages
    group admins in group pages
- If a template view is deleted from the site,
then all the locked copies it has will remain locked.

behatnotneeded

Change-Id: I47e98cf8413da5ba6d5dea0e29dddbd80c481317