Bug 1880803: update view_rows_columns when copying
we need to replace the method we use for checking if the
view_rows_columns records need to be updated or not.
When copying a page this is not saved correctly.

This table, related to old view layouts, contains
the number of columns for each row.

If the view layout has been defined with manually
edited columns, then it won't have a layout id but
it will have the view_rows_columns records up to date

behatnotneeded

Change-Id: Ibebd7a715943315e50158404215046c0568c9626