Bug 1875995 Consolidating list-group styles
Reducing the amount of overrides by creating new class
.list-group-top-border for .list-groups that need the top border in
first-child.

Fixing this also brought up a lot of random inconsistencies so
there are fixes for small things, like taking out some h4/h3 in
scss to only call up the class name (getting ready for when we
consolidate headings and font sizes), using .list-group-item-heading
for headings that don't need heading font and using .title for
when we may need to theme as headings.

behatnotneeded

Change-Id: Ia7e91717247ab1d05cf38c08468995959974912a