mahara_behat.sh add util.php option to get behat data root
mahara_behat.sh needs the behat data root. Currently it obtains this value
by using grep to extract the value from the config.php.

But this does not work if the value is set using some php function.

Add an option to util.php to get the value (which is do the right thing
if the value is set with a function) and get rid of the grep.

Change-Id: I33064bc75c624badcf48f235b59d4d6e09e3874f