Bug 1897831 Slant edge of Maroon theme with skins
When skin is applied with background images, maroon slants messing
up the images, so have changed the height and negative bottom
position of pageheader to make it overlap main content properly.
Negative top position also applied to footer to make it overlap
main content. Instead of using skew for the pageheader, have
changed to using rotate, and this required an additional wrapper in
the templates to hide the sides or it will have horizontal scroll.

behatnotneeded

Change-Id: I9e90f3c6ae621811c9b20a0ae30087e3f5651795