Bug 1862370: New submissions plugin
This plugin adds a comfortable management of current and already
released group submissions by providing a sort and filterable table in
its own group tab including direct access to the portfolios resp. their
archived version.

Additional metadata is also provided together with the possibility to
release the portfolio or to assign an evaluator and an evaluation result.

behatnotneeded: Behat tests to follow

Change-Id: I702322e3af92bd84a333edd028fcf8611e942bed