Bug 1906596: Added a dev-mode docker compose file.
Additionally added some instructions on how to use it.
Fix some issues in the default .env file and the base mahara image so
that it works with the instructions.

Updated Makefile to use a single make target to stand up an environment.

Updated docker-compose.dev.yaml to work with local volumes.

Removed volume names for now. Revisit this when making it work for
multiple environments.

Added 'make up' and 'make down' for use after the environment has been
built.

Added 'up', 'down' and 'new-dev-environment' targets to the 'help'
target.

Refactored the 'help' target message for readability.

Change-Id: I8f9a42f890b770eb2af9994658f7e09c31f644a7