Bug 1895066
Draft peer assessment needs additional indicator besides colour

Added a 'draft' string to "view" and used that in the template.
Updated the 'draft' string in "blocktype.peerassessment/peerassessment"
to 'savedraft' to prevent confusion.
Added 'pubstatus' to the item if it is private and set this to Draft.
Added pubstatus to the template if set to private. Style for this
replicates the private status on Bug 1912414.

Change-Id: I9cb0b20f9e00536e3162c3f83583b4877a6913a9