Bug 1915469 Flex titles and buttons
Buttons in list-group-item that float right and stick to the top
has various breakpoints that do different things, so redoing them
to use flex in a more consistent way and reduce the need for many
overrides in the theme.

Expected behaviour:
In list-group-items, edit and delete buttons etc stick to the top
right corner of that list-group item at all browser widths, even
if title wraps.

Actual behaviour:
The buttons stick at the top when in desktop width, then unstick
on tablet view, and some align to left at some breakpoint, and when
title wraps it would unstick at desktop width, or neglect the
user icon if there is one to the left of the title.

Also added new variable for user-icons too easily change out the
border-radius.

* Refactored selectors for Behat tests to locate the button area.
* Restore the metadata div if it is a private message (Annotations)
* Restore comment-text to commentlist.tpl for behat.
* Journals defaults to auto focus on first Journal title for multi-journal users.

behatnotneeded

Change-Id: I6a51e5e9597c3138867a588552bfc38141be3db3