Bug 1730263 - Multiple sources for names in persondata
A partial export (just selected pages) does not include the persondata
tags. When these are not present the first/last names are taken from the
author tag.  I've added a 'source' parameter to the firstname/lastname
fields that can be checked later in the process when the Import form is
being presented.  If the source is the author tags we force the Ignore
option on the decision field values.

The Bug referenced 'personal information' so I have included the email
field.

Change-Id: Ifa173a88f497be994c5ab93e26376463442ab24b