Bug 1916833 - Annotation can be updated on a page
Config gear icon removed from the Annotation blocks.
In order to still initially create these when added we needed to check
if the Block was a new instance or not. To achieve that we are now
passing the block instance into the has_instance_config() method.

BlockInstance has been extended and will set a 'new' parameter if the ID
in the constructor is 0.

AuthPlugin* classes also have the has_instance_config(). These do not
need the BlockInstance passed in.

Change-Id: Ifac69fc1d48b45c079b6f66e610776cead3dee63