Bug 1903601: LTI - Check for "sub" in the auth_remote_user
The user_id we get when creating users on cron has the form [GUID]_[ID].
This does not match the user_id we usually get which is of the form
[ID].  On other LTI calls the [GUID]_[ID] value is sent in the "sub"
parameter in the data from the JWT packet.

This will add a check for the "sub" value into the
module_lti_advantage_ensure_user_exists() request.

Change-Id: If3939f36fbfb6f690f23d7ff60d006461e7837ea
Signed-off-by: Gold <gold@catalyst.net.nz>