Bug 1887748: A WCAG 2.1 AA collection  navigation
Test:
- ✅ check that the forwards/backwards buttons work
- ✅ check that the text page number is correct

Use <button> instead of <span>
Multiple <nav> within a page
<nav> tags are used for major block of navigation links such as for the
main navigation of a page screen readers and other accessibility tools
may hide/skip the rendering of this content.
[reference] https://www.w3schools.com/tags/tag_nav.asp
When using multiple navigations within a page, their aria-label
should help to differentiate that their content is not the same.

<span> shouldn't be used in place of <buttons> as they're there
for when we need an inline wrapper for styling. We shouldn't be using
them for clickable items which should be buttons.

Fixed the styleguide code too for collection nav dropdowns

Change-Id: Ia6bd48e8744bf1765e5cba774d4cd89d89dc35e6