Bug 1918193: Add a function to allow buttons to redirect
Please test on child patches by interacting with
links that have been coverted to buttons using the 
'Enter' and 'Space' keys.

This function looks for the all the data-url attributes
in the page and sets that url to be used for the target
when (buttons) are interacted with.

This is especially needed as previously <a> tag href
attributes do not get overwritten by javascript
but data-url (js attributes do) so we need a way of
telling the page to find them again.

Change-Id: I20edecb98f9056f97af89db0f76f476ab8e902f3