Bug 800511 directing 'Contact us' emails
If a person is logged in and is a member of
an institution, their 'Contact us' email is
sent to their institution admin.

If they do not have an institution admin,
it is sent to the site admin.

If they are not a member of an institution
or they are logged out, it is sent to
the site admin.

Email message footer has been updated, and
institution admins can adjust their
'Contact us' email settings.

Change-Id: Ia55feb53a7001d6bb3069464c082d29209874c53