Bug 1953411: Creating a support admin role
This role sits between institution staff and institution administrator,
giving someone staff permissions as well as the permission to masquerade
as anybody in the institution who is an institution member or staff, i.e.
someone with fewer permissions than themselves.

Institution support administrators cannot suspend or delete an
institution member.

Change-Id: Icabbcc239af74b190ec7c1df493470d7884bdb6e
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>