Bug 1951103: Quick filter not displayed in submission
This happens when a language has an apostrophe in the language
string. Due to the JavaScript in the template file, the
apostrophe would have needed to be escaped as well. The fix
changes the code instead.

Change-Id: I7d1c51d896750f966b5894d60d8c9413b95c26a8
(cherry picked from commit bee7c07ae0509c4a74f9b7818cd4a8c53b373cee)