Bug 1959860: Plugin info link needs to work as button also
Before we were checking only for a <A> parent but now
we check for either an <A> or <Button> parent that has the data
attribute we need.

This patch makes it possible to click the 'Info' button at
Admin menu -> Extensions -> Plugin administration -> courseinfo
and have the modal appear with the additional information.

Change-Id: I3a94c21d84933662f67f7e35a2a1887e845693d9
Signed-off-by: Robert Lyon <robertl@catalyst.net.nz>