Bug 1964428: Preserve sign-off block config in template copies
Prevent deleting and editing the block instance.

Note for personal collections:
As the functionality for collection templates only
exist on an institution level at this point, you will
need to set 'Template' to 'Yes' for all pages in a
personal collection to ensure copies will preserve
the sign-off block config in all the pages in copies
of the collection.

The view_instruction_lock table only watches for individual
pages with the 'Template' switch set to 'Yes'/

Change-Id: Ie18073a5bb4d6c822f15ca023ae9a0eab4876033