Bug 1971898: Pull libraries via Composer
This initial test removes the recaptcha library, adds this to the
composer.json, installs it via make init composer, adds a behat test for
it, and adds sanity checks to init.php to check for the new vendor
directory.

To test, install a clean site and point a browser at it.

I've added cwegans/composer-patches and a patch on recaptcha that
removes the examples and tests dir.

Configure the site to use recaptcha under Site options > Security.
Enable on the registration and contact forms and insert the site and
secret keys.

For full testing you will want to create test keys for your
development environment at https://www.google.com/recaptcha/

For minimal testing (verifying that the library is loading and for
testing with behat) you can use the public test keys here:
https://developers.google.com/recaptcha/docs/faq#id-like-to-run-automated-tests-with-recaptcha.-what-should-i-do

While logged out browse to the Register and Contact pages.

Change-Id: I44b5669781b0c5ea6a85d5fbdb2c3e0617be4099