Bug 1976408: Resolving SQL errors on PHP8.1 during site install
While standing up on an empty DB the PHP8.1 version showed a decent
number of SQL errors while the PHP7.4 version did not. These errors are
due to code talking to the DB before the site is installed and the
absence of the tables throwing the errors.

Note the PHP 7.4 and 8.1 sites are served from the same codebase with
the same config and DB.

I've adjusted the get_config('installed') check to also make sure the
config table exists in the database as well.

Signed-off-by: Gold <gold@catalyst.net.nz>
Change-Id: I68efbf9744761fefcf6212b4786190165d271c99