Bug 1976416: PHP 8 Compatibility: api
ssl_*() functions:

https://www.php.net/manual/en/function.openssl-seal.php
https://www.php.net/manual/en/function.openssl-open.php

cipher_algo (param 5) is no longer an optional parameter. The default is RC4
but is considered insecure.

Someone that knows more about what is going on here should chip in.  Can we
just switch out the cipher_algo to something more reliable or does this
interact with outside sources that also need to be updated?

Change-Id: I7aa784d3bcf4fba110ab75f082ee74c6ae95eca5