Use the existing $noedit flag instead of the custom $noaddcopy
This is because $noedit is only assigned on group pages. This will
cause this flag to always be null, and therefore false and the
hardcoded true for the customisation will make it that the 'add and
copy' buttons appear.

We can instead use the existing $noedit flag only to simplify the logic.