Bug 1976432: PHP 8 Compatibility: webservice
Still a WIP as there is a question around session_get_instance(). The
function doesn't exist. Odds are that this code has never been hit as
there are no bug reports. This leads me to think that this should be a
stand alone bug report.  Will leave this as a WIP until a decision is
made on that though.

https://git.mahara.org/mahara/mahara/-/blob/main/htdocs/webservice/lib.php#L1251

Signed-off-by: Gold <gold@catalyst.net.nz>
Change-Id: If0c2e22e850e77536a358dbd71dc68f5ddd10f00