Bug 1980021: session_get_instance() does not exist
Digging into this I found that this was added in commit
e099440c5839516a61863809683ffcfbd92f9740 and at that point the function
wasn't present either. A full search of the git history shows only that
commit that includes the string so I'm thinking it is a safe bet that
this code has never been executed.

Removing the call and surrounding code.

Signed-off-by: Gold <gold@catalyst.net.nz>
Change-Id: I82c8a837e3d2f3359839830148c75321c7fead6d