Bug 1996475: Submit a copy of a Portfolio
Done:
Copy creation for View and Collection
Identification as a submission in View Index
Link back to source Portfolio
Link back to source if available on the Collection edit settings.
Link back to source if available on the View edit settings.
Link back to source if available on the view/index.php Portfolios summary page.
Title is modified for submissions.
Portfolio comments are copied.
Submissions that have been released are still hidden as submissions
A submission can be "Promoted" to being a normal portfolio

Smart Evidence is copied
Annotations are copied
Feedback on Annotations are copied
Assessments on Annotations are copied

Progress Completion is copied
Sign-off toggle and status are copied

This is working when submitting to Groups.
This is working when submitting to LTI Advantage.
This is in testing for submitting to LTI 1.1.

Signed-off-by: Gold <gold@catalyst.net.nz>
Change-Id: I3d07bab6bd81e86f1cb9e1193af00820a65455db